refactor: Modal confirm style
