+++ components/modal/style/index.tsx
@@ -1,9 +1,11 @@
 import type React from 'react';
-import { clearFix, genFocusStyle, resetComponent } from '../../style';
+
+import { genFocusStyle, resetComponent } from '../../style';
 import { initFadeMotion, initZoomMotion } from '../../style/motion';
+import type { GlobalToken } from '../../theme';
 import type { AliasToken, FullToken, GenerateStyle } from '../../theme/internal';
 import { genComponentStyleHook, mergeToken } from '../../theme/internal';
-import type { TokenWithCommonCls } from '../../theme/util/genComponentStyleHook';
+import type { GenStyleFn, TokenWithCommonCls } from '../../theme/util/genComponentStyleHook';
 
 /** Component only token. Which will handle additional calculation of alias token */
 export interface ComponentToken {
@@ -290,94 +292,6 @@ const genModalStyle: GenerateStyle<ModalToken> = (token) => {
   ];
 };
 
-const genModalConfirmStyle: GenerateStyle<ModalToken> = (token) => {
-  const { componentCls } = token;
-  const confirmComponentCls = `${componentCls}-confirm`;
-
-  return {
-    [confirmComponentCls]: {
-      '&-rtl': {
-        direction: 'rtl',
-      },
-      [`${token.antCls}-modal-header`]: {
-        display: 'none',
-      },
-      [`${confirmComponentCls}-body-wrapper`]: {
-        ...clearFix(),
-      },
-      [`${confirmComponentCls}-body`]: {
-        display: 'flex',
-        flexWrap: 'wrap',
-        alignItems: 'center',
-
-        [`${confirmComponentCls}-title`]: {
-          flex: '0 0 100%',
-          display: 'block',
-          // create BFC to avoid
-          // https://user-images.githubusercontent.com/507615/37702510-ba844e06-2d2d-11e8-9b67-8e19be57f445.png
-          overflow: 'hidden',
-          color: token.colorTextHeading,
-          fontWeight: token.fontWeightStrong,
-          fontSize: token.titleFontSize,
-          lineHeight: token.titleLineHeight,
-
-          [`+ ${confirmComponentCls}-content`]: {
-            marginBlockStart: token.marginXS,
-            flexBasis: '100%',
-            maxWidth: `calc(100% - ${token.modalConfirmIconSize + token.marginSM}px)`,
-          },
-        },
-
-        [`${confirmComponentCls}-content`]: {
-          color: token.colorText,
-          fontSize: token.fontSize,
-        },
-
-        [`> ${token.iconCls}`]: {
-          flex: 'none',
-          marginInlineEnd: token.marginSM,
-          fontSize: token.modalConfirmIconSize,
-
-          [`+ ${confirmComponentCls}-title`]: {
-            flex: 1,
-          },
-
-          // `content` after `icon` should set marginLeft
-          [`+ ${confirmComponentCls}-title + ${confirmComponentCls}-content`]: {
-            marginInlineStart: token.modalConfirmIconSize + token.marginSM,
-          },
-        },
-      },
-      [`${confirmComponentCls}-btns`]: {
-        textAlign: 'end',
-        marginTop: token.marginSM,
-
-        [`${token.antCls}-btn + ${token.antCls}-btn`]: {
-          marginBottom: 0,
-          marginInlineStart: token.marginXS,
-        },
-      },
-    },
-
-    [`${confirmComponentCls}-error ${confirmComponentCls}-body > ${token.iconCls}`]: {
-      color: token.colorError,
-    },
-
-    [`${confirmComponentCls}-warning ${confirmComponentCls}-body > ${token.iconCls},
-        ${confirmComponentCls}-confirm ${confirmComponentCls}-body > ${token.iconCls}`]: {
-      color: token.colorWarning,
-    },
-
-    [`${confirmComponentCls}-info ${confirmComponentCls}-body > ${token.iconCls}`]: {
-      color: token.colorInfo,
-    },
-
-    [`${confirmComponentCls}-success ${confirmComponentCls}-body > ${token.iconCls}`]: {
-      color: token.colorSuccess,
-    },
-  };
-};
-
 const genRTLStyle: GenerateStyle<ModalToken> = (token) => {
   const { componentCls } = token;
   return {
@@ -393,95 +307,53 @@ const genRTLStyle: GenerateStyle<ModalToken> = (token) => {
   };
 };
 
-const genWireframeStyle: GenerateStyle<ModalToken> = (token) => {
-  const { componentCls, antCls } = token;
-  const confirmComponentCls = `${componentCls}-confirm`;
+// ============================== Export ==============================
+export const prepareToken: (token: Parameters<GenStyleFn<'Modal'>>[0]) => ModalToken = (token) => {
+  const headerPaddingVertical = token.padding;
+  const headerFontSize = token.fontSizeHeading5;
+  const headerLineHeight = token.lineHeightHeading5;
 
-  return {
-    [componentCls]: {
-      [`${componentCls}-content`]: {
-        padding: 0,
-      },
+  const modalToken = mergeToken<ModalToken>(token, {
+    modalBodyPadding: token.paddingLG,
+    modalHeaderPadding: `${headerPaddingVertical}px ${token.paddingLG}px`,
+    modalHeaderBorderWidth: token.lineWidth,
+    modalHeaderBorderStyle: token.lineType,
+    modalHeaderBorderColorSplit: token.colorSplit,
+    modalHeaderHeight: headerLineHeight * headerFontSize + headerPaddingVertical * 2,
+    modalFooterBorderColorSplit: token.colorSplit,
+    modalFooterBorderStyle: token.lineType,
+    modalFooterPaddingVertical: token.paddingXS,
+    modalFooterPaddingHorizontal: token.padding,
+    modalFooterBorderWidth: token.lineWidth,
+    modalIconHoverColor: token.colorIconHover,
+    modalCloseIconColor: token.colorIcon,
+    modalCloseBtnSize: token.fontSize * token.lineHeight,
+    modalConfirmIconSize: token.fontSize * token.lineHeight,
+  });
 
-      [`${componentCls}-header`]: {
-        padding: token.modalHeaderPadding,
-        borderBottom: `${token.modalHeaderBorderWidth}px ${token.modalHeaderBorderStyle} ${token.modalHeaderBorderColorSplit}`,
-        marginBottom: 0,
-      },
-
-      [`${componentCls}-body`]: {
-        padding: token.modalBodyPadding,
-      },
-
-      [`${componentCls}-footer`]: {
-        padding: `${token.modalFooterPaddingVertical}px ${token.modalFooterPaddingHorizontal}px`,
-        borderTop: `${token.modalFooterBorderWidth}px ${token.modalFooterBorderStyle} ${token.modalFooterBorderColorSplit}`,
-        borderRadius: `0 0 ${token.borderRadiusLG}px ${token.borderRadiusLG}px`,
-        marginTop: 0,
-      },
-    },
-
-    [confirmComponentCls]: {
-      [`${antCls}-modal-body`]: {
-        padding: `${token.padding * 2}px ${token.padding * 2}px ${token.paddingLG}px`,
-      },
-      [`${confirmComponentCls}-body`]: {
-        [`> ${token.iconCls}`]: {
-          marginInlineEnd: token.margin,
-
-          // `content` after `icon` should set marginLeft
-          [`+ ${confirmComponentCls}-title + ${confirmComponentCls}-content`]: {
-            marginInlineStart: token.modalConfirmIconSize + token.margin,
-          },
-        },
-      },
-      [`${confirmComponentCls}-btns`]: {
-        marginTop: token.marginLG,
-      },
-    },
-  };
+  return modalToken;
 };
 
-// ============================== Export ==============================
+export const prepareComponentToken = (token: GlobalToken) => ({
+  footerBg: 'transparent',
+  headerBg: token.colorBgElevated,
+  titleLineHeight: token.lineHeightHeading5,
+  titleFontSize: token.fontSizeHeading5,
+  contentBg: token.colorBgElevated,
+  titleColor: token.colorTextHeading,
+});
+
 export default genComponentStyleHook(
   'Modal',
   (token) => {
-    const headerPaddingVertical = token.padding;
-    const headerFontSize = token.fontSizeHeading5;
-    const headerLineHeight = token.lineHeightHeading5;
+    const modalToken = prepareToken(token);
 
-    const modalToken = mergeToken<ModalToken>(token, {
-      modalBodyPadding: token.paddingLG,
-      modalHeaderPadding: `${headerPaddingVertical}px ${token.paddingLG}px`,
-      modalHeaderBorderWidth: token.lineWidth,
-      modalHeaderBorderStyle: token.lineType,
-      modalHeaderBorderColorSplit: token.colorSplit,
-      modalHeaderHeight: headerLineHeight * headerFontSize + headerPaddingVertical * 2,
-      modalFooterBorderColorSplit: token.colorSplit,
-      modalFooterBorderStyle: token.lineType,
-      modalFooterPaddingVertical: token.paddingXS,
-      modalFooterPaddingHorizontal: token.padding,
-      modalFooterBorderWidth: token.lineWidth,
-      modalIconHoverColor: token.colorIconHover,
-      modalCloseIconColor: token.colorIcon,
-      modalCloseBtnSize: token.fontSize * token.lineHeight,
-      modalConfirmIconSize: token.fontSize * token.lineHeight,
-    });
     return [
       genModalStyle(modalToken),
-      genModalConfirmStyle(modalToken),
       genRTLStyle(modalToken),
       genModalMaskStyle(modalToken),
-      token.wireframe && genWireframeStyle(modalToken),
       initZoomMotion(modalToken, 'zoom'),
     ];
   },
-  (token) => ({
-    footerBg: 'transparent',
-    headerBg: token.colorBgElevated,
-    titleLineHeight: token.lineHeightHeading5,
-    titleFontSize: token.fontSizeHeading5,
-    contentBg: token.colorBgElevated,
-    titleColor: token.colorTextHeading,
-  }),
+  prepareComponentToken,
 );
