+++ components/modal/style/wireframeCmp.tsx
@@ -1,0 +1,69 @@
+// Style as confirm component
+import { prepareComponentToken, prepareToken, type ModalToken } from '.';
+import { genSubStyleComponent, type GenerateStyle } from '../../theme/internal';
+
+// ============================= Confirm ==============================
+
+const genWireframeStyle: GenerateStyle<ModalToken> = (token) => {
+  const { componentCls, antCls } = token;
+  const confirmComponentCls = `${componentCls}-confirm`;
+
+  return {
+    [componentCls]: {
+      [`${componentCls}-content`]: {
+        padding: 0,
+      },
+
+      [`${componentCls}-header`]: {
+        padding: token.modalHeaderPadding,
+        borderBottom: `${token.modalHeaderBorderWidth}px ${token.modalHeaderBorderStyle} ${token.modalHeaderBorderColorSplit}`,
+        marginBottom: 0,
+      },
+
+      [`${componentCls}-body`]: {
+        padding: token.modalBodyPadding,
+      },
+
+      [`${componentCls}-footer`]: {
+        padding: `${token.modalFooterPaddingVertical}px ${token.modalFooterPaddingHorizontal}px`,
+        borderTop: `${token.modalFooterBorderWidth}px ${token.modalFooterBorderStyle} ${token.modalFooterBorderColorSplit}`,
+        borderRadius: `0 0 ${token.borderRadiusLG}px ${token.borderRadiusLG}px`,
+        marginTop: 0,
+      },
+    },
+
+    [confirmComponentCls]: {
+      [`${antCls}-modal-body`]: {
+        padding: `${token.padding * 2}px ${token.padding * 2}px ${token.paddingLG}px`,
+      },
+      [`${confirmComponentCls}-body`]: {
+        [`> ${token.iconCls}`]: {
+          marginInlineEnd: token.margin,
+
+          // `content` after `icon` should set marginLeft
+          [`+ ${confirmComponentCls}-title + ${confirmComponentCls}-content`]: {
+            marginInlineStart: token.modalConfirmIconSize + token.margin,
+          },
+        },
+      },
+      [`${confirmComponentCls}-btns`]: {
+        marginTop: token.marginLG,
+      },
+    },
+  };
+};
+
+// ============================== Export ==============================
+export default genSubStyleComponent(
+  ['Modal', 'wireframe'],
+  (token) => {
+    const modalToken = prepareToken(token);
+
+    return [token.wireframe && genWireframeStyle(modalToken)];
+  },
+  prepareComponentToken,
+  {
+    // Should after other modal style
+    order: -997,
+  },
+);
